Handle varying lengths of subtitles and translations
Description:
This commit modifies the `merge_srt_and_translation` and `create_translated_srt` functions to correctly handle subtitles and translations of varying lengths. Previously, these functions assumed that each subtitle in the `.srt` file and the corresponding translation in the `.txt` file always occupy the same number of lines. However, this is not the case when a subtitle or its translation consists of multiple sentences, which may be on separate lines.

Changes made:
1. In the `merge_srt_and_translation` function, we now read the original text from the `.srt` file until we encounter a line that is either a number or an empty line (indicating the start of the next subtitle). We also read the translated text from the `.txt` file until we encounter a line that is a number (indicating the start of the next translation).

2. Similarly, in the `create_translated_srt` function, we now read the translated text from the `.txt` file until we encounter a line that is a number (indicating the start of the next translation).

These changes ensure that our functions correctly handle subtitles and translations that consist of multiple sentences on separate lines.